Moved handlers to ResponseProvider
